send "chat user" message when a user changes its name
When a user of a chat we are in changes its name, send a "chat user"
message to the client with the following format:

"chat: user: <account_id> <chat_room> <new_name> <old_name> <status>"

In this case, status is "rename". New_name can be a new alias or a new
name.

Signed-off-by: hwipl <33433250+hwipl@users.noreply.github.com>